+++ tests/ci/unit_tests/deploy_me/application_steps/__init__.py
@@ -1,11 +1,6 @@
-#  Copyright (c) 2021 Food-X Technologies
+#  Copyright (c) 2022 Food-X Technologies
 #
 #  This file is part of foodx_devops_tools.
 #
 #  You should have received a copy of the MIT License along with
 #  foodx_devops_tools. If not, see <https://opensource.org/licenses/MIT>.
-
-"""FoodX core deployment utility."""
-
-from ._main import deploy_me  # noqa: F401
-from ._state import ExitState  # noqa: F401
