+++ tests/ci/unit_tests/deploy_me/deployment/test_deploy_step.py
@@ -11,7 +11,7 @@ import logging
 import pytest
 
 from foodx_devops_tools._to import StructuredTo
-from foodx_devops_tools.deploy_me._deployment import _deploy_step
+from foodx_devops_tools.deploy_me.application_steps import deploy_step
 
 
 @pytest.fixture()
@@ -25,7 +25,8 @@ def mock_run_puff(mock_async_method):
 @pytest.fixture()
 def mock_rg_deploy(mock_async_method):
     this_mock = mock_async_method(
-        "foodx_devops_tools.deploy_me._deployment.deploy_resource_group"
+        "foodx_devops_tools.deploy_me.application_steps._deploy"
+        ".deploy_resource_group"
     )
     return this_mock
 
@@ -33,7 +34,8 @@ def mock_rg_deploy(mock_async_method):
 @pytest.fixture()
 def mock_login(mock_async_method):
     mock_async_method(
-        "foodx_devops_tools.deploy_me._deployment.login_service_principal"
+        "foodx_devops_tools.deploy_me.application_steps"
+        "._deploy.login_service_principal"
     )
 
 
@@ -46,7 +48,7 @@ async def test_clean(
     mock_run_puff,
     mock_verify_puff_target,
 ):
-    await _deploy_step(**mock_deploystep_context)
+    await deploy_step(**mock_deploystep_context)
 
     mock_rg_deploy.assert_called_once()
 
@@ -65,7 +67,7 @@ async def test_default_override_parameters(
     this_context = copy.deepcopy(mock_deploystep_context)
     this_context["this_step"].static_secrets = False
 
-    await _deploy_step(**this_context)
+    await deploy_step(**this_context)
 
     expected_defaults = default_override_parameters(
         mock_deploystep_context["deployment_data"]
@@ -121,7 +123,7 @@ async def test_secrets_enabled(
         }
     )
 
-    await _deploy_step(**this_context)
+    await deploy_step(**this_context)
 
     mock_rg_deploy.assert_called_once_with(
         mocker.ANY,
@@ -145,7 +147,7 @@ async def test_step_skip(
             frame="f1", application="a1", step="other_step"
         )
 
-        await _deploy_step(**mock_deploystep_context)
+        await deploy_step(**mock_deploystep_context)
 
         mock_rg_deploy.assert_not_called()
